pyproject.toml
[project]
name = "SearchInVideos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "google-cloud-aiplatform>=1.76.0",
    "google-generativeai>=0.8.3",
    "moviepy>=2.1.1",
    "pydub>=0.25.1",
    "streamlit>=1.41.1",
]
